20220717-SeparateButtonActionListenerFromWritingListPanel
버튼 액션리스너 리스트 패널로 부터 분리
