+++ input.csv
@@ -1,7 +1,2 @@
-23423,2342,PUBLISHED
-여기는 공게시판3,3414,PUBLISHED
-여기는 공설게시판,공설게시판,PUBLISHED
-한번더?  아니 두병더,두병더!,PUBLISHED
-지금이거는 아난오겟지,안나오고 전에꺼 나오겟지,PUBLISHED
-null,null,PUBLISHED
-yo inu,yoyo,PUBLISHED
+이랗ㄱ게 쓴다 ,그냥 이렇게,DELETED
+dlfjg,ffgew2g,PUBLISHED
