+++ private.csv
@@ -1,11 +1,3 @@
-23423,2342,PUBLISHED
-2여기는 사설게시판,3414,PUBLISHED
-여기는 사설이얌,사섫,PUBLISHED
-사설이라니깐 개색기야,왜 말을안들어,PUBLISHED
-null,null,PUBLISHED
-dk tsibnal fgqadfdioswgjeiowgjwe,	r2hrs wasted fuck man ,, wtf   ffgjhitjiwj 
-but iam happy dont worry  i am goo man yiou all good man
-
-
-
-si bal,PUBLISHED
+이렇게[ 부ㅠㄴ리 해도 되나요 ,dfㅎㄹㅎㄴㅎㅇ,PUBLISHED
+아니 이렇게 분리해줘도 되는거 맞냐고,대체 맞냐니깐
+,PUBLISHED
